chore(menu-full-screen): convert unit tests to RTL
